Dramatically refactor async design.
A few fun notes:

* lib/runner - An abstraction wrapping our use of async.series, which:
  * takes a heterogeneous array of (1) user-defined test hooks/functions 
and (2) arrays of such functions. These can be arbitrarily deeply nested
which I did both to separate the arrangement of the actions from their
execution and to pave the way for a runner that supports arbitrarily
nested tests (as opposed to just single-layer xUnit style tests). 
  * build a results object which has both a global flag to indicate 
whether ANYTHING failed (in which case the whole test is a failure) as
well as a "nestedPassing" collection of switches localized to the 
nesting. For instance, if a global beforeAll fails, ALL the tests should
be considered failures. If a test module's beforeAll fails, all the
tests _defined inside that module_ should fail. And so on. This was a 
super tricky concept to encode, and I ultimately settled for making it
mutable (which means test parallelizing or reordering won't be easily 
possible)

* lib/build-test-actions - build the nested arrays of test functions
that are fed to the runner

* lib/build-test-helper - builds the test helper; if one is provided
by the user, its file will be available for any error messages of global
failures so users can more easily track the error down

* lib/user-function-async-wrapper-factory - given the options and log,
returns a function which can be used to wrap any user function (e.g.
a hook or a test function) and make it quack as if it's asynchronous.
For genuinely async functions, it will enforce the test's timeout for 
them as well. Additionally, it'll log any failures against the mutable
`results` object instantiated by the runner, separating "error" messages
from "failure" messages (e.g. if the test fails, print the `not ok` even
if the test itself has no errror)