Dramatically increase the complexity of everything
This was certainly one way to make sure the right hooks are firing when
we want. 

Particularly nasty is the use of `distanceFromTest` as a crude 
instrument to determine whether an afterEach is at a deeper depth than
an earlier error (to avoid running them unnecessarily). Just ugly as 
hell.