Separate the Results plugin from the Tap13 logging
This was pretty challenging and required us to do a few things:

* Every test action now comes with an incremented id for each test run
(userFunction, test, suite), which is used by plugins to look up in 
their own stores for unique stuff
* the results plugin now quacks like all the other ones, returning a
normal plugin object and not requiring the log arg. that's nice; weirdly
the behavior for test & suite scoped results-building is identical once
you remove the logging concern
* the core plugin wrapper will now pass {error: someError} to user 
functions that errored and whose error was swallowed when building 
results. This was needed to have any way to pass beforeAll/afterAll 
errors forward to the tap13 reporter plugin