Refactor all non-functional state thru one point
There was way too much state floating around previously through 3 
different "store" modules, each with their own initialization and root
data structure. The potential for leaking state across multiple runs
was far too great to leave this as-was, especially given the need for
additional state to be stored.

This means that all non-functional state is always reset when the
test runner is kicked off. It also means that users will not be 
permitted to register plugins except when the runner deems it 
appropriate, which is a little bit inconvenient, but we've given users
3 ways of configuring plugins, so hopefully that's enough.