Update configuration to handle multiple file/glob patterns
This allows `criteria()` to accept either a single string or an array of test locator globs. It also accepts a second `userOptions` argument which is used to provide `userOptions.example` as a globally applied `name` filter.

The return value is an array of objects, each with a `.file` property and, possibly, a `.name` and `.lineNumber` property, each of which is an array of filters that should be applied to that file.